requirements.txt
python-dotenv
langchain-nvidia-ai-endpoints
langchain_openai
langchain
fastapi
uvicorn
twilio
singleton-decorator
kagglehub
selenium
langchain
langchain-nvidia-ai-endpoints
pillow
transformers
flax
jax
webdriver-manager
easyocr
sentencepiece